Refactor subprocessing
